requirements.txt
# Core ML dependencies
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost

# Database and cloud storage
pymongo
google-cloud-storage
google-cloud-logging
google-cloud-secret-manager

# Web framework
fastapi
uvicorn
jinja2
python-multipart

# MLOps and utilities
evidently==0.2.8
dill
PyYAML
neuro_mf
from_root

# Development and testing
pytest
pytest-asyncio
httpx

# Local development
-e . 